# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management API",
        "description": "API documentation for Agile Project Management System - PM → Project → Sprint → Task workflow",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tasks/{task}/sub-tasks": {
            "get": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Get all subtasks for a task",
                "description": "Get all subtasks for a task",
                "operationId": "87c0442a65cc0003d464fc75b54c55ab",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subtasks"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Create a new subtask",
                "description": "Create a new subtask",
                "operationId": "9a0b698c7e51f554069627dcf69c9731",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "tag": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "SubTask created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/{task}/sub-tasks/{subTask}": {
            "put": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Update subtask",
                "description": "Update subtask",
                "operationId": "1c549cfcc997bac25318b80d3ce31182",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subTask",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "tag": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_completed": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SubTask updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "SubTask not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Delete subtask",
                "description": "Delete subtask",
                "operationId": "21d5311b4362891f53aa663dfc2c8225",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subTask",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SubTask deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "SubTask not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your access token from login. Example: 1|abc123xyz...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}